Allow image loading to BYTEA Greenplum column
If the user defines the external table with an image data column of
bytea or int[] type, the query will automatically change its output.

Also we add another metadata column, to keep track of the number of
images and their dimesionality, for example, the 4th column should now
have '{100,256,128}' for a row containing 100 images of width 256 and
height 128 pixels. This is to facilitate reading the data from the bytea
column (last column).

Authored-by: Oliver Albertini <oalbertini@pivotal.io>