pxf cli: make reset more verbose
Often, especially in CI, pxf reset command fails silently. The reason is
usually related to an unset PXF_CONF variable, but still, it would be
nice to get some output.

Instead of redirecting the output of doStatus, this commit sets an
environment variable that allows for silencing of the typical doStatus
output, which the user doesn't need when trying to reset the cluster.
This will allow other errors to surface.

Authored-by: Oliver Albertini <oalbertini@vmware.com>