Fix bug when modifying PXF's classpath in automation (#861)
PXF's automation test framework includes a helper method
(`addPathToPxfClassPath`) that modifies pxf-env.sh to set
`PXF_LOADER_PATH`. The problem is that the methods reads the content of
pxf-env.sh from $PXF_BASE/conf on the local host which may not be the
same as the file on a multinode cluster. Since the remote configuration
is copied down from the remote cluster to a temporary working directory,
the helper method should load pxf-env.sh from this location.

Additionally, this commit increases the size of PXF heap for multinode
tests. The original heap size of 512 MB was too small for some of our
tests but this wasn't noticed until now because the automation framework
modified pxf-env.sh without preserving any changes made to it during the
installation and setup of PXF on the multinode cluster.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>