Disable actuator shutdown endpoint
The recommended way for shutting down a Spring Boot application is to
send the JVM SIGTERM. This commit disables the shutdown endpoint since
we can always assume that PXF is being shutdown locally and not
remotely. Additionally, the status command of the PXF CLI now only
checks if the PXF JVM process exists; if the PXF webapp fails to load
for some reason, the JVM exits. Finally, the scripts dependency on curl
can be removed.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>