Update jackson-databind to 2.13.4.1 (#886)
This commit updates the jackson-databind version to 2.13.4.1 in the automation framework as a precautionary measure to mitigate CVE-2022-42003.

PXF pulls in jackson-databind through Springboot v2.5.12. The CVE in question is related to UNWRAP_SINGLE_VALUE_ARRAYS which is disabled by default in jackson-databind and stays disabled in Springboot v2.5.12. We remove explicit declarations of these dependencies in the subproject build.gradle files.