Pass `ENVIRONMENT` var to `Promote PXF-GP5 and PXF-GP6 Artifacts` task.
Making it easier to work on the release pipeline in a development
environment, the `Promote PXF-GP5 and PXF-GP6 Artifacts` release
pipeline task and underlying `promote_pxf_artifacts` task yml and
corresponding bash script will be passed the `ENVIRONMENT` variable
containing `prod` (production release) or `dev` (development release).

This change was specifically introduced to control the Tagging and
Bumping Version Number operations.

Setting the `ENVIRONMENT` variable to `prod` will perform the Tagging
and Bumping Version Number operations. Otherwise, these operations are
skipped.

Additionally, for dev release pipelines set via the Makefile, set the
pipeline variable `ud/pxf/dev/git-remote-ssh-url` to `noop`. This
value is assigned to `GIT_REMOTE_URL`. It is unused in the development
release pipeline environment. This assignment remains in the Makefile
to allow the possibility for developers to perform tagging operations
on an alternate non-production git repo.

Update README.md to include reference to the Makefile target
`dev-release` to depkloy a developer version of the release pipeline.