ignore parquetWriteListReadWithHive test (#913)
For parquetWriteListReadWithHive test, we use HIVE to read out Parquet Data we write using pxf. This test can pass on HDP2 and HDP3, but it fails on CDH. CDH is using HIVE1.1, which doesn't support Parquet Date. Parquet Date is supported by Hive1.2+. So for now, we decide to temporarily ignore this test.