Separate logging and user output in pxf-cli
The cluster-level utility for managing the PXF service was using the
gplog package to both log to a file and display messages to the user of
the utility. This PR separates these two concerns. Unfortunately gplog
does not currently have a way to completely disable writing log messages
to the terminal.

To disable logging to the terminal, the terminal log level is set to
LOGERROR and all logging statements are at LOGINFO or lower. Messages
for the user are written to the terminal using the fmt package.

The commit also adds a new multinode test job to the dev pipeline that
only tests the PXF CLI.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>