Fix pxf-dev-base build configration (#931)
According to the Cloud Build documentation [1], the substitutions field
is a map of string keys to string values. Unfortunately in YAML, an
unquoted value of 1.19 is parsed as a number, causing the build to fail
with

> json: cannot unmarshal number into Go value of type string

This commit quotes the substitution value of 1.19 to force it to be
parsed as a string.

Additionally, this commit installs python2 and python2-pip in the GPDB7
RHEL8 pxf-dev-base image. These packages used to be included in the base
container image but have since been removed. PXF still needs python2 to
run tinc tests.

[1]: https://cloud.google.com/build/docs/build-config-file-schema

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>