Remove unnecessary container images (#970)
Ubuntu 18.04 will not be a supported platform for GP7 and GP RelEng has
deleted the upstream image gpdb7-ubuntu18.04-test. This commit removes
the PXF dev base container build for GP7 Ubuntu 18.04. It also updates
the EL8 container image for GP6 to use Rocky 8, same as the container
image for GP7.

Authored-By: Bradford D. Boyle <bradfordb@vmware.com>