Add Make variable to conditionally include check-creds (#975)
When setting a Concourse pipeline, the set-pipeline command can take an
optional flag named --check-creds that will validate credential
variables against the credential manager. Lately this validation can
take an extended period of time which can slow down the development
cycle when iterating on CI pipelines. Rather than remove the flag
altogether and risk missing an incorrect credential in a production
pipeline, this commit adds a new Make varialble named CHECK_CREDS which
defaults to true. If this variable is true, then the --check-creds flag
to set-pipeline will be included.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>