fix rpmrebuild-rocky8 container image (#1000)
There is a compatibility issue between the version of coreutils
(v8.30-15) included in rockylinux:8 and the version of Concourse
(v5.7.1) used for PXF's CI. This causes the command `ls <some-dir>` to
fail with the error message

```sh
ls: cannot access '<some-dir>': Operation not permitted
```

The same comand in the same container image running on Concourse 7.8.3
works successfully.

The Greenplum release engineering team previsouly encountered this same
issue and resolved it by downgrading coreutils-single to v8.30-8. This
version is no longer available from upstream repos for downgrading in
the rpmrebuild images. Instead, this commit switches the base image from
upstream rockylinux:8 to gpdb6-rocky8-build which is maintained by the
Greenplum release engineering team and includes the older version of
coreutils-single. For consistency, this commit also changes upstream
centos:7 to gpdb6-centos7-build.

Other changes:

* add upstream base images to triggers
* update container image diagram

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>