Refactor cloudbuild (#1008)
This commit creates a script that downloads the specified version of maven-3. You can also specify `latest` to grab the latest maven-3 version.

This commit also refactors the cloudbuild to pull the maven tarball from GCS instead of from upstream Apache. This is to prevent future potential pipeline issues. Additionally, we do not need the latest build of maven to run automation tests.