ci: fix issue in automation Makefile (#1021)
For all of the protocols other than file, PXF requires
$(PROTOCOL)-site.xml to configure the particular protocol (e.g., AWS
credentials for s3 protocol). Rather then repeating the copying of the
required template from $PXF_HOME, the automation Makefile has the copy
command at the start of the shell script for the sync_cloud_config
target.The configuration for the file protocol is done via pxf-site.xml
instead of file-site.xml. This causes the copy command to attempt to
copy a non-existent file-site.xml. This commit adds a check for the
protocl before attempting to copy.

The reason the failed copy doesn't fail the build in CI is that a
multiline shell command is used in the recipe for sync_cloud_configs and
this shell command does not include errexit. The shell keeps processing
the multiline command despite the failing copy command and returns the
last exist code which is 0 so make considers the command a success. This
commit enables errexit to cause the shell to immediately exit if any
command exits with a non-zero status.

Authored-by: Bradford D. Boyle <bradfordb@vmware.com>