Add pxf.service.kerberos.ticket-renew-window option to pxf-site.xml (#1018)
PXF currently renews the kerberos ticket only when at least 80% of the ticket lifespan has passed. 

This PR introduces a new property `pxf.service.kerberos.ticket-renew-window` that allows users to change the threshold at which we renew the ticket. The default value is 0.8f to maintain backwards compatibility. This value can be changed by adding the property into the pxf-site.xml file and can be different for each server config. For example, to renew the ticket after 30% of the ticket lifespan has passed, set the following: 
```
    <property>
        <name>pxf.service.kerberos.ticket-renew-window</name>
        <value>0.3</value>
    </property>
```