Install go to /opt/pxf/go in the docker images (#1024)
This commit changes the install location in the Docker images from /usr/local/go to /opt/pxf/go.
This change is necessitated by the concourse version currently being used (v5.7.1) and can be reverted upon a concourse version upgrade. 

It also adds a sha25sum check to make sure we are downloading the correct binary.

It also updates the concourse dev pipeline to include PXF_DEV_IMAGE_TAG as an option to make for faster testing with different images.