Change dataproc script to use golang-petname (#1067)
This commit updates the package used in the dataproc script that generates a random cluster name.

This change is needed because the google/cloud-sdk:slim image used to spin up the dataproc clusters has been bumped to Debian 12 (Bookworm) and comes with Python 3.11. Python 3.11 introduces a concept called externally managed environments (see PEP 668) which now distinguishes between user-installed packages and system packages.

Instead of using a python package, we can directly download the golang-petname package, a Debian supported package with the same functionality.