Reduce flakiness by waiting for HDFS file to exist (#1073)
This commit uses awaitility to wait for the HDFS file to exist before we run the SQL query. Metadata about a file is written to the Namenode after the file is written and closed. This is when the file is considered accessible, so waiting for the file to exist should ensure that we do not start reading while data is still being written to the files used by testing.