Update job trigger for certification pipeline (#1084)
The pxf-build pipeline pushes release tags to Github after the release artifacts have been successfully pushed to required buckets.

The certification pipeline's pxf_src resource updates when there is a new release tag on Github. As such, the trigger for the certification pipeline should be on the pxf_src and not on the pxf artifacts. Changing the trigger would ensure that the certification pipeline is being properly kicked off: when both the new release tag is present on Github and the artifacts are present in the required buckets.

[skip ci]