Refactor position tracking for multi-line JSON files (#1100) (#1114)
JsonRecordReader was refactored in commit 515235f (PR #879) to internally use the LineRecordReader when handling multi-line JSON files.

Previously, we were calculating the position inside JsonRecordReader by finding the length of bytes from the current index to the end of the current line that was read in by the underlying LineRecordReader. This causes performance issues for files with a lot of data condensed into very few lines.

Now, instead of calculating the number of bytes from the current index in the line to the end of the line, we calculate the length of the object we successfully read. Additionally, we take advantage of the underlying LineRecordReader to determine the true starting position of a split.