Adjusted PXF error handling and determining client disconnects (#1111)
Spring MVC 5.3.33+ wraps ClientAbortException into its own AsyncRequestNotUsableException when client disconnect happens on async thread, so we need to adjust how we determine this scenario.

This PR also changes exception handling when the response is committed. It used to set an error on the response, but this was causing an IllegalStateException from Tomcat. Now we just re-throw the original exception to make Tomcat terminate the client connection abnormally without sending the closing 0-length chunk, thus informing the client that there was a processing error.