Modify build pipeline to make it easier to set release pipelines (#1113)
This commit removes the build make target in favor of 2 new make targets: release and main.

As we no longer expect releases to be done off main, make main spins up the current pxf-build pipeline. However, shipit files for the main branch point to ${BUCKET}/shipit/noop and should effectively be ignored.

The release make target is to be used for all releases going forward and expects the variable RELEASE_BRANCH to be set. Release branches are expected in the form branch-<major.minor>.x. We expect these release pipelines to be long standing and so a release for 6.10.99 would be expected to use the same pipeline as 6.10.4.

Running make release RELEASE_BRANCH=branch-6.10.x would do the following:

create a pipeline with the name pxf-build-branch-6.10.x that points to the branch-6.10.x branch for deployment
any shipit file would live in ${BUCKET}/shipit/branch-6.10.x