Add support for timestamptz to JDBC profile (#1104)
This PR adds read and write support for `timestamptz` for the JDBC profile for any profile/connector that supports the JDBC 4.2 API. As of this PR, the Hive JDBC connector does not fully support the JDBC 4.2 API and thus PXF does not support `timestamptz` when trying to connect to Hive using JDBC. 

This PR also extends functionality added in commit e1b148a. Wide date range `date` and `timestamp` with eras will continue to silently fail and return incorrect data if the `date_wide_range` option is turned off. Wide date range `date` and `timestamp` without eras now uses `date_wide_range` option as a hint for better performance.  If a user does not provide the `date_wide_range` option, instead of erroring out, the JDBC profile will fallback to the `date_wide_range` formatter and attempt to parse the data a second time. In such a scenario, the JDBC resolver will log a warning to suggest turning on `date_wide_range` to get better performance.