Add release branch script (#1126)
This commit adds a script that allows developers to easily create a release branch for a particular major.minor version, update the version file and tag any associated images needed for the release pipeline. This script also prints out the necessary make command to set the release pipeline for the branch.