Add timeout to test-file-tpl job (#1133)
This commit adds a 2 hr timeout to the test-file-tpl.yml template. It is possible for the job to run for hours on end, RCA unconfirmed. To prevent the job from running forever, add a timeout to the job. As it usually takes 30-40 minutes for the job to run, set the timeout for 2 hrs to give some buffer.